+++ src/main/java/com/gonghak98/v2/report/domain/abeek/dto/CheckResult.java
@@ -6,7 +6,7 @@ import com.gonghak98.v2.report.domain.counting.AreaCreditSummary;
 import java.util.Map;
 
 public record CheckResult(Map<AreaType, Boolean> passResults,
-                          Map<Integer, NonPassMessage> nonPassResults,
+                          Map<Long, NonPassMessage> nonPassResults,
                           Map<AreaType, AreaCreditSummary> creditSummaries) {
 
 }
